+++ dyninstAPI/src/inst-aarch64.C
@@ -28,51 +28,12 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA
  */
 
-#include "codegen/emitters/aarch64/EmitterAarch64Dyn.h"
-#include "codegen/emitters/aarch64/EmitterAarch64Stat.h"
-#include "Instruction.h"
-#include "Register.h"
-#include "registerSpace/RegisterConversion.h"
 #include "codegen/codegen.h"
-#include "registerSpace/registerSpace.h"
+#include "dyntypes.h"
+#include "emitter.h"
 
-using codeGenASTPtr = Dyninst::DyninstAPI::codeGenASTPtr;
-using operandAST = Dyninst::DyninstAPI::operandAST;
-
-void emitLoadPreviousStackFrameRegister(Address register_num, Register dest,
-                                        codeGen &gen, int /*size*/) {
+void emitLoadPreviousStackFrameRegister(Dyninst::Address register_num,
+                                        Dyninst::Register dest, codeGen &gen,
+                                        int /*size*/) {
   gen.codeEmitter()->emitLoadOrigRegister(register_num, dest, gen);
-}
-
-// First AST node: target of the call
-// Second AST node: source of the call
-// This can handle indirect control transfers as well
-bool AddressSpace::getDynamicCallSiteArgs(InstructionAPI::Instruction i,
-                                          Address addr, std::vector<codeGenASTPtr> &args) {
-  namespace di = Dyninst::InstructionAPI;
-
-  auto cft = i.getControlFlowTarget();
-  auto target_reg = boost::dynamic_pointer_cast<di::RegisterAST>(cft);
-  if (!target_reg)
-    return false;
-  auto branch_target = convertRegID(target_reg);
-
-  if (branch_target == registerSpace::ignored)
-    return false;
-
-  // jumping to Xn (BLR Xn)
-  args.push_back(operandAST::origRegister((void *)(long)branch_target));
-  args.push_back(operandAST::Constant((void *)addr));
-
-  return true;
-}
-
-Emitter *AddressSpace::getEmitter() {
-  static Dyninst::DyninstAPI::EmitterAarch64Stat emitter64Stat;
-  static Dyninst::DyninstAPI::EmitterAarch64Dyn emitter64Dyn;
-
-  if (proc())
-    return &emitter64Dyn;
-
-  return &emitter64Stat;
 }
